Add code that passes the smoke test for conversation endpoint
