+++ backend/send_me/modules/conversations/models.py
@@ -5,7 +5,6 @@ from sqlalchemy import DateTime, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from send_me.database.models import Base
-from send_me.modules.users.models import User
 
 """
 This class represents a Conversation in the database.
@@ -26,8 +25,8 @@ class Conversation(Base):
         DateTime(timezone=True), nullable=False, default=datetime.now
     )
 
-    newest_message = relationship("Message", back_populates="conversation")
-    users = relationship("UserConversations", back_populates="conversation")
+    newest_message = relationship("messages", back_populates="conversation")
+    users = relationship("user_conversations", back_populates="conversation")
 
 
 class UserConversations(Base):
@@ -38,7 +37,7 @@ class UserConversations(Base):
     conversation_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("conversations.id"))
     read: Mapped[bool] = mapped_column(default=False)
 
-    user = relationship("User", back_populates="conversations")
+    user = relationship("users", back_populates="conversations")
 
 
 class Message(Base):
@@ -49,8 +48,8 @@ class Message(Base):
     conversation_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("conversations.id"))
     content: Mapped[str]
 
-    sender = relationship(User, back_populates="mesages")
-    conversation = relationship(Conversation, back_populates="messages")
+    sender = relationship("users", back_populates="mesages")
+    conversation = relationship("conversations", back_populates="messages")
 
     created_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), nullable=False, default=datetime.now
